admin-dashboard user list and order list

+++ BE/BE/Repositories/Interfaces/IUsersRepository.cs
@@ -7,9 +7,9 @@ namespace BE.Repositories.Interfaces
     public interface IUsersRepository
     {
         Task<IEnumerable<Users>> GetAllAsync();
-        Task<Users?> GetByIdAsync(int id);
+        Task<Users?> GetByIdAsync(long id);
         Task<Users> AddAsync(Users user);
-        Task<Users?> UpdateAsync(int id, Users user);
-        Task<bool> DeleteAsync(int id);
+        Task<Users?> UpdateAsync(long id, Users user);
+        Task<bool> DeleteAsync(long id);
     }
 }
